LPS-192654 Move to right package
